Use general category to determine zero width combining characters
Previously, the canonical combining class was used to determine
which characters are zero width combining characters. This had
two problems:
- it classified spacing marks (category Mc) as zero width
- it classified enclosing marks (category Me) as normal characters

Fix by using the general category to generate a table of combining
characters that have zero width. Characters with a general category
of Mn or Me are included in this table. Characters with a general
category of Mc are not included, and so use the default width of 1.