Switch to using importlib.metadata
pkg_resources is a little heavyweight, and has some rather glaring
shortcomings for checking the metadata of an install Python package.
importlib.metadata is contained in the standard library (as of 3.8), and
provides a much nicer interface. Switch to using it, falling back to
importlib_metadata if the Python version is too old.